Move the `Filesystem cache` dashboard into a separate file
The dashboard covers every filesystem cache metric, so it changes on
every metric addition. Keeping it in `StorageSystemDashboards.cpp` would
make that shared file a conflict hotspot, so move the charts into
`StorageSystemDashboardsFilesystemCache.cpp` behind a
`getFilesystemCacheDashboards` accessor, following the same pattern as
`getDistributedCacheDashboards`. The shared file keeps only a four line
hook.
